feat: implements user to audit zipcode

+++ src/main/java/com/igorbavand/ceptracker/infrastructure/ZipCodeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ZipCodeRepository extends JpaRepository<ZipCode, Integer> {
+public interface ZipCodeRepository extends JpaRepository<ZipCode, Long> {
 
     @Query("SELECT z FROM ZipCode z WHERE REPLACE(z.zipCode, '-', '') = REPLACE(:zipCode, '-', '')")
     List<ZipCode> findAllByZipCodeIgnoringDash(String zipCode);
